Removes misleading text from walkable row when user is near the stop
Fixes #935 - The walk indicator can be wrong when you're standing at the bus stop

+++ OBAKit/Categories/NSDate+DateTools.h
@@ -25,7 +25,7 @@
 
  @return representiation of minutes
  */
-- (double)minutesUntil;
+- (NSUInteger)minutesUntil;
 
 - (BOOL)isToday;
 - (BOOL)isTomorrow;
